refactor!: move extensions to their own crate
