+++ .github/workflows/ci.yml
@@ -20,10 +20,6 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: Swatinem/rust-cache@v2
-      - name: Install GDAL
-        run: |
-          sudo apt-get update
-          sudo apt-get install libgdal-dev
       - name: Test
         run: cargo test --lib
   lint:
@@ -31,10 +27,6 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: Swatinem/rust-cache@v2
-      - name: Install GDAL
-        run: |
-          sudo apt-get update
-          sudo apt-get install libgdal-dev
       - name: Fmt
         run: cargo fmt
       - name: Clippy
@@ -50,10 +42,6 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: Swatinem/rust-cache@v2
-      - name: Install GDAL
-        run: |
-          sudo apt-get update
-          sudo apt-get install libgdal-dev
       - name: Install check-all-features
         run: cargo install cargo-all-features
       - name: Check
@@ -64,10 +52,6 @@ jobs:
       - uses: actions/checkout@v4
       - uses: dtolnay/rust-toolchain@nightly
       - uses: Swatinem/rust-cache@v2
-      - name: Install GDAL
-        run: |
-          sudo apt-get update
-          sudo apt-get install libgdal-dev
       - name: Check
         run: cargo check --all --all-features
   doc:
@@ -77,10 +61,6 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: Swatinem/rust-cache@v2
-      - name: Install GDAL
-        run: |
-          sudo apt-get update
-          sudo apt-get install libgdal-dev
       - name: Doc
         run: cargo doc --all-features
   validate-stac-server:
