+++ .github/workflows/cli.yml
@@ -26,7 +26,7 @@ env:
 jobs:
   test-cli:
     strategy:
-      matrix: 
+      matrix:
         os:
           - macos-latest
           - windows-latest
@@ -37,17 +37,6 @@ jobs:
         uses: Swatinem/rust-cache@v2
       - name: Test
         run: cargo test --no-default-features -F pgstac -F python
-  test-cli-with-gdal:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - uses: Swatinem/rust-cache@v2
-      - name: Install GDAL
-        run: |
-          sudo apt-get update
-          sudo apt-get install libgdal-dev
-      - name: Test
-        run: cargo test --no-default-features -F gdal
   test-cli-with-duckdb:
     runs-on: ubuntu-latest
     steps:
