+++ cli/README.md
@@ -10,17 +10,14 @@ Command Line Interface (CLI) for [STAC](https://stacspec.org/), named `stacrs`.
 
 ## Installation
 
-```sh
-cargo install stac-cli
+```shell
+python -m pip install stacrs-cli
 ```
 
 Or:
 
-```shell
-# NOTE: The version from PyPI does not include GDAL or DuckDB support. If you
-# need to use these features, install via `cargo install` (GDAL is enabled by
-# default) or `cargo install -F duckdb` (DuckDB is not).
-python -m pip install stacrs-cli
+```sh
+cargo install stac-cli
 ```
 
 Then:
@@ -44,15 +41,14 @@ Use the `--help` flag to see all available options for the CLI and the subcomman
 
 ## Features
 
-This crate has five features, three of them on by default:
+This crate has features:
 
 - `duckdb`: experimental support for querying [stac-geoparquet](https://github.com/stac-utils/stac-geoparquet) files using [DuckDB](https://duckdb.org/)
-- `gdal`: read geospatial data from rasters (enabled by default)
 - `geoparquet`: read and write [stac-geoparquet](https://github.com/stac-utils/stac-geoparquet) (enabled by default)
 - `pgstac`: enable a [pgstac](https://github.com/stac-utils/pgstac) backend for `stacrs serve` (enabled by default)
 - `python`: create an entrypoint that can be called from Python (used to enable `python -m pip install stacrs-cli`)
 
-If you don't want to use GDAL or any of the other default features:
+If you don't want to use any of the default features:
 
 ```shell
 cargo install stac-cli --no-default-features
