+++ duckdb/Cargo.toml
@@ -1,25 +1,26 @@
 [package]
 name = "stac-duckdb"
-version = "0.0.2"
-authors = ["Pete Gadomski <pete.gadomski@gmail.com>"]
-edition = "2021"
 description = "Experimental client for querying stac-geoparquet using DuckDB"
-homepage = "https://github.com/stac-utils/stac-rs"
-repository = "https://github.com/stac-utils/stac-rs"
-license = "MIT OR Apache-2.0"
+version = "0.0.2"
 keywords = ["geospatial", "stac", "metadata", "geo", "raster"]
-categories = ["science", "data-structures"]
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+categories.workspace = true
+rust-version.workspace = true
 
 [dependencies]
-arrow = { workspace = true }
-duckdb = { workspace = true }
-libduckdb-sys = { workspace = true }
-geoarrow = { workspace = true }
-parquet = { workspace = true }
+arrow.workspace = true
+duckdb.workspace = true
+libduckdb-sys.workspace = true
+geoarrow.workspace = true
+parquet.workspace = true
 stac = { workspace = true, features = ["geoarrow"] }
-stac-api = { workspace = true }
-thiserror = { workspace = true }
+stac-api.workspace = true
+thiserror.workspace = true
 
 [dev-dependencies]
 duckdb-test = { path = "duckdb-test" }
-geo = { workspace = true }
+geo.workspace = true
