+++ extensions/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "stac-extensions"
+description = "Manage STAC extensions (https://stac-extensions.github.io/)"
+version = "0.1.0"
+keywords = ["geospatial", "stac", "extensions"]
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+categories.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+geojson.workspace = true
+stac.workspace = true
+serde.workspace = true
+serde_json.workspace = true
