+++ extensions/examples
@@ -1,0 +1,1 @@
+../spec-examples/v1.1.0
