+++ pgstac/Cargo.toml
@@ -1,26 +1,27 @@
 [package]
 name = "pgstac"
-version = "0.2.1"
-authors = ["Pete Gadomski <pete.gadomski@gmail.com>"]
-edition = "2021"
 description = "Rust interface for pgstac"
-homepage = "https://github.com/stac-utils/stac-rs"
-repository = "https://github.com/stac-utils/stac-rs"
-license = "MIT OR Apache-2.0"
+version = "0.2.1"
 keywords = ["geospatial", "stac", "metadata", "raster", "database"]
 categories = ["database", "data-structures", "science"]
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+rust-version.workspace = true
 
 [features]
 tls = ["dep:rustls", "dep:tokio-postgres-rustls", "dep:webpki-roots"]
 
 [dependencies]
-geojson = { workspace = true }
+geojson.workspace = true
 rustls = { workspace = true, features = ["ring", "std"], optional = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-stac = { workspace = true }
-stac-api = { workspace = true }
-thiserror = { workspace = true }
+serde.workspace = true
+serde_json.workspace = true
+stac.workspace = true
+stac-api.workspace = true
+thiserror.workspace = true
 tokio-postgres = { workspace = true, features = ["with-serde_json-1"] }
 tokio-postgres-rustls = { workspace = true, optional = true }
 webpki-roots = { workspace = true, optional = true }
@@ -28,7 +29,7 @@ webpki-roots = { workspace = true, optional = true }
 [dev-dependencies]
 pgstac-test = { path = "pgstac-test" }
 tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
-tokio-test = { workspace = true }
+tokio-test.workspace = true
 
 [package.metadata.docs.rs]
 all-features = true
