+++ pgstac/pgstac-test/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "pgstac-test"
 version = "0.0.0"
-edition = "2021"
+edition.workspace = true
 publish = false
 
 [lib]
@@ -10,6 +10,6 @@ test = false
 doctest = false
 
 [dependencies]
-quote = { workspace = true }
+quote.workspace = true
 syn = { workspace = true, features = ["full", "extra-traits"] }
-tokio-postgres = { workspace = true }
+tokio-postgres.workspace = true
