+++ python/Cargo.toml
@@ -1,11 +1,13 @@
 [package]
 name = "python"
 description = "Python bindings for stac-rs"
-license = "MIT OR Apache-2.0"
-repository = "https://github.com/stac-utils/stac-rs"
-homepage = "https://github.com/stac-utils/stac-rs"
 version = "0.1.3"
-edition = "2021"
+authors.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+edition.workspace = true
+rust-version.workspace = true
 publish = false
 
 [lib]
@@ -16,11 +18,11 @@ crate-type = ["cdylib"]
 duckdb = { workspace = true, features = [
     "bundled",
 ] } # we don't use it directly, but we need to ensure it's bundled
-geojson = { workspace = true }
+geojson.workspace = true
 pyo3 = { workspace = true, features = ["extension-module"] }
-pythonize = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
+pythonize.workspace = true
+serde.workspace = true
+serde_json.workspace = true
 stac = { workspace = true, features = [
     "geoparquet-compression",
     "object-store-all",
@@ -28,5 +30,5 @@ stac = { workspace = true, features = [
     "validate-blocking",
 ] }
 stac-api = { workspace = true, features = ["client"] }
-stac-duckdb = { workspace = true }
+stac-duckdb.workspace = true
 tokio = { workspace = true, features = ["rt"] }
