+++ server/Cargo.toml
@@ -1,15 +1,15 @@
 [package]
 name = "stac-server"
-version = "0.3.1"
-edition = "2021"
-authors = ["Pete Gadomski <pete.gadomski@gmail.com>"]
 description = "SpatioTemporal Asset Catalog (STAC) API server"
-homepage = "https://github.com/stac-utils/stac-rs"
-repository = "https://github.com/stac-utils/stac-rs"
-license = "MIT OR Apache-2.0"
-keywords = ["geospatial", "stac", "metadata", "geo", "raster"]
+version = "0.3.1"
+keywords = ["geospatial", "stac", "metadata", "geo", "server"]
 categories = ["science", "data-structures"]
-rust-version = "1.75"
+edition.workspace = true
+authors.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+rust-version.workspace = true
 
 [features]
 axum = ["dep:axum", "dep:bytes", "dep:mime", "dep:tower-http"]
@@ -20,24 +20,24 @@ axum = { workspace = true, optional = true }
 bb8 = { workspace = true, optional = true }
 bb8-postgres = { workspace = true, optional = true }
 bytes = { workspace = true, optional = true }
-http = { workspace = true }
+http.workspace = true
 mime = { workspace = true, optional = true }
 pgstac = { workspace = true, features = ["tls"], optional = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-serde_urlencoded = { workspace = true }
-stac = { workspace = true }
+serde.workspace = true
+serde_json.workspace = true
+serde_urlencoded.workspace = true
+stac.workspace = true
 stac-api = { workspace = true, features = ["geo"] }
-thiserror = { workspace = true }
+thiserror.workspace = true
 tokio-postgres = { workspace = true, optional = true }
 tower-http = { workspace = true, features = ["cors"], optional = true }
-url = { workspace = true }
+url.workspace = true
 
 [dev-dependencies]
-serde_json = { workspace = true }
+serde_json.workspace = true
 stac = { workspace = true, features = ["validate"] }
 tokio = { workspace = true, features = ["macros"] }
-tokio-test = { workspace = true }
+tokio-test.workspace = true
 tower = { workspace = true, features = ["util"] }
 
 [package.metadata.docs.rs]
